feat/order-service

+++ location-service/src/main/java/at/ac/fhcampuswien/se/group1/locationservice/repository/LocationRepository.java
@@ -4,6 +4,8 @@ import at.ac.fhcampuswien.se.group1.locationservice.model.Location;
 import org.springframework.data.mongodb.repository.MongoRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface LocationRepository extends MongoRepository<Location, Integer> {
+public interface LocationRepository extends MongoRepository<Location, BigInteger> {
 }
